client/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sbsync-client"
version = "0.1.0"
description = "Obsidian vault git pull client utility"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.46",
    "prometheus-client>=0.23.1",
    "python-dotenv>=1.2.1",
    "schedule>=1.2.0",
]

[project.scripts]
sbsync-client = "src.main:main"

[dependency-groups]
dev = [
    "pytest>=7.4.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
